Change Auth Page separate sign up page into phoneCheck and sign change sign route to auth add another sign route
