+++ src/components/auth/phoneCheck/index.ts
@@ -1,0 +1,1 @@
+export { default } from './phoneCheck.component';
