+++ src/components/auth/welcome/index.ts
