+++ src/components/auth/welcome/welcome.module.scss
