--- src/container/sign/index.ts
@@ -1,1 +1,0 @@
-export { default } from './sign.container';
