--- src/pages/sign/index.ts
@@ -1,1 +1,0 @@
-export { default } from './sign.page';
