--- src/pages/sign/sign.page.tsx
@@ -1,10 +1,0 @@
-import * as React from 'react';
-import SignContianer from '../../container/sign';
-
-const SignPage: React.SFC<{}> = () => {
-  return (
-    <SignContianer />
-  );
-}
-
-export default SignPage;
